+++ test/Cabazure.Messaging.EventHub.Tests/EventHubModelsGenerator.cs
@@ -1,33 +1,20 @@
-using AutoFixture.Kernel;
+using AutoFixture;
 using Azure.Messaging.EventHubs;
+using Cabazure.Test.Customizations;
 
 namespace Cabazure.Messaging.EventHub.Tests;
 
-public class EventHubModelsGenerator : ISpecimenBuilder
+public class EventHubModelsGenerator : TypeCustomization<EventData>
 {
-    /// <inheritdoc/>
-    public object Create(object request, ISpecimenContext context)
+    public EventHubModelsGenerator()
+        : base(fixture => EventHubsModelFactory.EventData(
+            eventBody: fixture.Create<BinaryData>(),
+            properties: fixture.Create<Dictionary<string, object>>(),
+            systemProperties: fixture.Create<Dictionary<string, object>>(),
+            partitionKey: fixture.Create<string>(),
+            sequenceNumber: fixture.Create<long>(),
+            offsetString: fixture.Create<string>(),
+            enqueuedTime: fixture.Create<DateTimeOffset>()))
     {
-        if (IsRequestFor<EventData>(request))
-        {
-            return EventHubsModelFactory.EventData(
-                eventBody: context.Create<BinaryData>(),
-                properties: context.Create<Dictionary<string, object>>(),
-                systemProperties: context.Create<Dictionary<string, object>>(),
-                partitionKey: context.Create<string>(),
-                sequenceNumber: context.Create<long>(),
-                offsetString: context.Create<string>(),
-                enqueuedTime: context.Create<DateTimeOffset>());
-        }
-
-        return new NoSpecimen();
     }
-
-    private static bool IsRequestFor<T>(object request)
-        => request switch
-        {
-            Type type => type == typeof(T),
-            SeededRequest seededRequest => Equals(seededRequest.Request, typeof(T)),
-            _ => false,
-        };
 }